% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{BamToBed_CLI}
\alias{BamToBed_CLI}
\title{Constructor method for BamToBed_CLI}
\usage{
BamToBed_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath,
  outputFormat)
}
\arguments{
\item{inFilePath}{(inFileNames may also be specified otherwise they will be fetched by list.files)}
}
\description{
Constructor method for BamToBed_CLI
}

