% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{class}
\name{Bowtie2_CLI-class}
\alias{Bowtie2_CLI-class}
\title{Bowtie2_CLI}
\description{
Bowtie2_CLI
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{inFilePath \code{"character"}}
   \item{\code{slot2}:}{inFileNames \code{"character"}}
   \item{\code{slot3}:}{cliParams \code{"character"}}
   \item{\code{slot4}:}{outFilePath \code{"character"}}
   \item{\code{slot5}:}{outputFlag \code{"character"}}
   \item{\code{slot6}:}{bowtieIndexFilePath \code{"character"}}
   \item{\code{slot7}:}{matepairFileNames \code{"character"}}
 }
}

