% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{class}
\name{Bowtie2TophatIon_CLI-class}
\alias{Bowtie2TophatIon_CLI-class}
\title{Bowtie2TophatIon_CLI}
\description{
Bowtie2TophatIon_CLI
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{inFilePath \code{"character"}}
   \item{\code{slot2}:}{inFileNames \code{"character"}}
   \item{\code{slot3}:}{cliParams \code{"character"}}
   \item{\code{slot4}:}{outFilePath \code{"character"}}
   \item{\code{slot5}:}{outputFlag \code{"character"}}
   \item{\code{slot6}:}{bowtieIndexFilePath \code{"character"}}
   \item{\code{slot7}:}{matepairFileNames \code{"character"}}
   \item{\code{slot8}:}{outFileNames \code{"character"}}
   \item{\code{slot9}:}{pathToPicardTools \code{"character"}}
   \item{\code{slot9}:}{addNHTag \code{"character"}}
 }
}

