% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{Bowtie2TophatIon_CLI}
\alias{Bowtie2TophatIon_CLI}
\title{Constructor method for Bowtie2TophatIon_CLI}
\usage{
Bowtie2TophatIon_CLI(inFilePath, inFileNames, cliParams, outputFlag,
  outFilePath, bowtieIndexFilePath, matepairFileNames, outFileNames,
  pathToPicardTools)
}
\description{
Constructor method for Bowtie2TophatIon_CLI
}

