% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{class}
\name{Bowtie_CLI-class}
\alias{Bowtie_CLI-class}
\title{Bowtie_CLI}
\description{
Bowtie_CLI
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{inFilePath \code{"character"}}
   \item{\code{slot2}:}{inFileNames \code{"character"}}
   \item{\code{slot3}:}{cliParams \code{"character"}}
   \item{\code{slot4}:}{outFilePath \code{"character"}}
   \item{\code{slot5}:}{outputFlag \code{"character"}}
   \item{\code{slot6}:}{bowtieIndexFilePath \code{"character"}}
   \item{\code{slot7}:}{matepairFileNames \code{"character"}}
   \item{\code{slot8}:}{addNHTag \code{"logical"}}
   \item{\code{slot9}:}{outputPipeString \code{"character"}}
   \item{\code{slot10}:}{outputFormat \code{"character"}}
 }
}

