% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{Bowtie_CLI}
\alias{Bowtie_CLI}
\title{Constructor method for Bowtie_CLI}
\usage{
Bowtie_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath,
  bowtieIndexFilePath, matepairFileNames, addNHTag, outputPipeString,
  outputFormat)
}
\arguments{
\item{inFilePath}{(inFileNames may also be specified otherwise they will be fetched by list.files)}
}
\description{
Constructor method for Bowtie_CLI
}

