% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{CmdGenResult}
\alias{CmdGenResult}
\title{Constructor method for CmdGenResult}
\usage{
CmdGenResult(CLIApplication, OutResultReference, commands)
}
\value{
\code{"CmdGenResult"}
}

