% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{class}
\name{CmdGenResultExec-class}
\alias{CmdGenResultExec-class}
\title{CmdGenResultExec}
\section{Slots}{
 
 \describe{
   \item{\code{slot1}:}{cmdGenResult \code{"CmdGenResult"}}
   \item{\code{slot2}:}{execLog \code{"execLog"}}
 }
}

