% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{executeCommandResult,CmdGenResult-method}
\alias{executeCommandResult,CmdGenResult-method}
\title{Executes the commands of a executeCommandResult object}
\usage{
\S4method{executeCommandResult}{CmdGenResult}(object, testing = FALSE,
  useTee = TRUE)
}
\arguments{
\item{\code{"testing"}}{boolean testing if true - command is not executed}
}

