% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{class}
\name{FilesOutput-class}
\alias{FilesOutput-class}
\title{FilesOutput}
\description{
FilesOutput
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{inFilePath \code{"character"}}
 }
}

