% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{FilesOutput}
\alias{FilesOutput}
\title{Constructor method for FilesOutput}
\usage{
FilesOutput(outResultName)
}
\arguments{
\item{outResultName}{}
}
\value{
\code{"FilesOutput"}
}
\description{
Constructor method for FilesOutput
}

