% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{class}
\name{FoldersOutput-class}
\alias{FoldersOutput-class}
\title{FoldersOutput}
\description{
FoldersOutput
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{inFilePath \code{"character"}}
 }
}

