% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{FoldersOutput}
\alias{FoldersOutput}
\title{Constructor method for FoldersOutput}
\usage{
FoldersOutput(outResultName)
}
\arguments{
\item{outResultName}{}
}
\value{
\code{"FoldersOutput"}
}
\description{
Constructor method for FoldersOutput
}

