% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{generateCommandResult}
\alias{generateCommandResult}
\alias{generateCommandResult,BamToBed_CLI-method}
\alias{generateCommandResult,Bowtie2TophatIon_CLI-method}
\alias{generateCommandResult,Bowtie2_CLI-method}
\alias{generateCommandResult,Bowtie_CLI-method}
\alias{generateCommandResult,Cutadapt_CLI-method}
\alias{generateCommandResult,Genomecov_CLI-method}
\alias{generateCommandResult,HTSeqCount_CLI-method}
\alias{generateCommandResult,IntersectBed_CLI-method}
\alias{generateCommandResult,MergeBedFile_CLI-method}
\alias{generateCommandResult,MultiBamCov_CLI-method}
\alias{generateCommandResult,MultiIntersectBed_CLI-method}
\alias{generateCommandResult,MultiIntersectBed_perl_CLI-method}
\alias{generateCommandResult,RNAStar_CLI-method}
\alias{generateCommandResult,Samtools_CLI-method}
\alias{generateCommandResult,SortFile_CLI-method}
\alias{generateCommandResult,Tophat2_CLI-method}
\title{Generates the commands and possibly resulting files/folders}
\usage{
generateCommandResult(object)

\S4method{generateCommandResult}{Cutadapt_CLI}(object)

\S4method{generateCommandResult}{HTSeqCount_CLI}(object)

\S4method{generateCommandResult}{IntersectBed_CLI}(object)

\S4method{generateCommandResult}{MergeBedFile_CLI}(object)

\S4method{generateCommandResult}{MultiBamCov_CLI}(object)

\S4method{generateCommandResult}{Genomecov_CLI}(object)

\S4method{generateCommandResult}{Samtools_CLI}(object)

\S4method{generateCommandResult}{BamToBed_CLI}(object)

\S4method{generateCommandResult}{SortFile_CLI}(object)

\S4method{generateCommandResult}{Bowtie2_CLI}(object)

\S4method{generateCommandResult}{Bowtie2TophatIon_CLI}(object)

\S4method{generateCommandResult}{Tophat2_CLI}(object)

\S4method{generateCommandResult}{Bowtie_CLI}(object)

\S4method{generateCommandResult}{MultiIntersectBed_CLI}(object)

\S4method{generateCommandResult}{MultiIntersectBed_perl_CLI}(object)

\S4method{generateCommandResult}{RNAStar_CLI}(object)
}
\value{
CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult
}

