% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{class}
\name{Genomecov_CLI-class}
\alias{Genomecov_CLI-class}
\title{Genomecov_CLI}
\description{
Genomecov_CLI
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{inFilePath \code{"character"}}
   \item{\code{slot2}:}{inFileNames \code{"character"}}
   \item{\code{slot3}:}{cliParams \code{"character"}}
   \item{\code{slot4}:}{outFilePath \code{"character"}}
   \item{\code{slot5}:}{outputFlag \code{"character"}} In this class the flag may be bedgraph (not used but specifies filetype)
   \item{\code{slot6}:}{outFileName \code{"character"}}
   \item{\code{slot7}:}{inFileBam \code{"logical"}}
 }
}

