% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{Genomecov_CLI}
\alias{Genomecov_CLI}
\title{Constructor method for Genomecov_CLI}
\usage{
Genomecov_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath,
  outFileName, inFileBam)
}
\arguments{
\item{inFilePath}{only one filename allowed!}
}
\description{
Constructor method for Genomecov_CLI
}

