% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getAddNHTag,Bowtie_CLI-method}
\alias{getAddNHTag,Bowtie_CLI-method}
\title{Accessor getAddNHTag}
\usage{
\S4method{getAddNHTag}{Bowtie_CLI}(object)
}
\value{
addNHTag
}
\description{
Accessor getAddNHTag
}

