% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getAnnotationFileMB}
\alias{getAnnotationFileMB}
\title{Accessor annotationFileMB}
\usage{
getAnnotationFileMB(object)
}
\value{
annotationFileMB
}
\description{
Accessor annotationFileMB
}

