% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getAnnotationType}
\alias{getAnnotationType}
\title{Accessor annotationType}
\usage{
getAnnotationType(object)
}
\value{
annotationType
}
\description{
Accessor annotationType
}

