% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{getBowtieIndexFilePath,Bowtie_CLI-method}
\alias{getBowtieIndexFilePath,Bowtie_CLI-method}
\title{Accessor getBowtieIndexFilePath}
\usage{
\S4method{getBowtieIndexFilePath}{Bowtie_CLI}(object)
}
\value{
bowtieIndexFilePath
}
\description{
Accessor getBowtieIndexFilePath
}

