% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{getBowtieIndexFilePath,Tophat2_CLI-method}
\alias{getBowtieIndexFilePath,Tophat2_CLI-method}
\title{Accessor getBowtieIndexFilePath Tophat2}
\usage{
\S4method{getBowtieIndexFilePath}{Tophat2_CLI}(object)
}
\value{
bowtieIndexFilePath
}
\description{
Accessor getBowtieIndexFilePath Tophat2
}

