% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getBowtieIndexFilePath}
\alias{getBowtieIndexFilePath}
\title{Accessor getBowtieIndexFilePath}
\usage{
getBowtieIndexFilePath(object)
}
\value{
bowtieIndexFilePath
}
\description{
Accessor getBowtieIndexFilePath
}

