% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getCLIApplication}
\alias{getCLIApplication}
\title{Accessor CLIApplication}
\usage{
getCLIApplication(object)
}
\value{
CLIApplication
}
\description{
Accessor CLIApplication
}

