% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getCmdGenResult}
\alias{getCmdGenResult}
\title{Accessor getCmdGenResult}
\usage{
getCmdGenResult(object)
}
\value{
getCmdGenResult
}
\description{
Accessor getCmdGenResult
}

