% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getExecLogFormatted}
\alias{getExecLogFormatted}
\title{Accessor getExecLogFormatted}
\usage{
getExecLogFormatted(object)
}
\value{
execLog formatted
}
\description{
Accessor getExecLogFormatted
}

