% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getFilterConditions}
\alias{getFilterConditions}
\title{Accessor getFilterConditions}
\usage{
getFilterConditions(object)
}
\value{
filterConditions
}
\description{
Accessor getFilterConditions
}

