% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getGenomeIndexFilePath}
\alias{getGenomeIndexFilePath}
\title{Accessor genomeIndexFilePath}
\usage{
getGenomeIndexFilePath(object)
}
\value{
genomeIndexFilePath
}
\description{
Accessor genomeIndexFilePath
}

