% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getGTFFilePath}
\alias{getGTFFilePath}
\title{Accessor getGTFFilePath}
\usage{
getGTFFilePath(object)
}
\value{
gtfFilePath
}
\description{
Accessor getGTFFilePath
}

