% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getInFileNames}
\alias{getInFileNames}
\title{Accessor getInFileNames}
\usage{
getInFileNames(object)
}
\value{
inFileNames
}
\description{
Accessor getInFileNames
}

