% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{getMatepairFileNames,Bowtie_CLI-method}
\alias{getMatepairFileNames,Bowtie_CLI-method}
\title{Accessor getMatepairFileNames}
\usage{
\S4method{getMatepairFileNames}{Bowtie_CLI}(object)
}
\value{
matepairFileNames
}
\description{
Accessor getMatepairFileNames
}

