% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getMatepairFileNames}
\alias{getMatepairFileNames}
\title{Accessor getMatepairFileNames}
\usage{
getMatepairFileNames(object)
}
\value{
matepairFileNames
}
\description{
Accessor getMatepairFileNames
}

