% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getOutFileName,Genomecov_CLI-method}
\alias{getOutFileName,Genomecov_CLI-method}
\title{Accessor getOutFileName}
\usage{
\S4method{getOutFileName}{Genomecov_CLI}(object)
}
\value{
outFileName
}

