% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{getOutFileName,MultiIntersectBed_perl_CLI-method}
\alias{getOutFileName,MultiIntersectBed_perl_CLI-method}
\title{Accessor getOutFileName}
\usage{
\S4method{getOutFileName}{MultiIntersectBed_perl_CLI}(object)
}
\value{
outFileName
}
\description{
Accessor getOutFileName
}

