% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getOutFileName}
\alias{getOutFileName}
\title{Accessor getOutFileName}
\usage{
getOutFileName(object)
}
\value{
outFileName
}
\description{
Accessor getOutFileName
}

