% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{getOutputFormat,BamToBed_CLI-method}
\alias{getOutputFormat,BamToBed_CLI-method}
\title{Accessor getOutputFormat}
\usage{
\S4method{getOutputFormat}{BamToBed_CLI}(object)
}
\value{
outputFormat
}
\description{
Accessor getOutputFormat
}

