% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getOutputFormat,Bowtie_CLI-method}
\alias{getOutputFormat,Bowtie_CLI-method}
\title{Accessor getOutputFormat}
\usage{
\S4method{getOutputFormat}{Bowtie_CLI}(object)
}
\value{
outputFormat
}

