% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getOutputFormat}
\alias{getOutputFormat}
\title{Accessor getOutputFormat}
\usage{
getOutputFormat(object)
}
\value{
outputFormat
}
\description{
Accessor getOutputFormat
}

