% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getOutputPipeString}
\alias{getOutputPipeString}
\title{Accessor getOutputPipeString}
\usage{
getOutputPipeString(object)
}
\value{
outputPipeString
}
\description{
Accessor getOutputPipeString
}

