% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getOutResultName}
\alias{getOutResultName}
\title{Accessor getOutResultName}
\usage{
getOutResultName(object)
}
\value{
outResultName
}
\description{
Accessor getOutResultName
}

