% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getOutResultReference}
\alias{getOutResultReference}
\title{Accessor getOutResultReference}
\usage{
getOutResultReference(object)
}
\value{
OutResultReference
}
\description{
Accessor getOutResultReference
}

