% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getPathToPicardTools}
\alias{getPathToPicardTools}
\title{Accessor getPathToPicardTools}
\usage{
getPathToPicardTools(object)
}
\value{
pathToPicardTools
}
\description{
Accessor getPathToPicardTools
}

