% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getRequireStrandness}
\alias{getRequireStrandness}
\title{Accessor getRequireStrandness}
\usage{
getRequireStrandness(object)
}
\value{
requireStrandness
}
\description{
Accessor getRequireStrandness
}

