% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getSamtoolsApplication}
\alias{getSamtoolsApplication}
\title{Accessor getSamtoolsApplication}
\usage{
getSamtoolsApplication(object)
}
\value{
samtoolsApplication
}
\description{
Accessor getSamtoolsApplication
}

