% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getSortInFiles}
\alias{getSortInFiles}
\title{Accessor getSortInFiles}
\usage{
getSortInFiles(object)
}
\value{
sortInFiles
}
\description{
Accessor getSortInFiles
}

