% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getStrandColumn}
\alias{getStrandColumn}
\title{Accessor getStrandColumn}
\usage{
getStrandColumn(object)
}
\value{
strandColumn
}
\description{
Accessor getStrandColumn
}

