% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getWithinGroupTH}
\alias{getWithinGroupTH}
\title{Accessor getWithinGroupTH}
\usage{
getWithinGroupTH(object)
}
\value{
withinGroupTH
}
\description{
Accessor getWithinGroupTH
}

