% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{HTSeqCount_CLI}
\alias{HTSeqCount_CLI}
\title{Constructor method for HTSeqCount_CLI}
\usage{
HTSeqCount_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath,
  gffFile)
}
\description{
Constructor method for HTSeqCount_CLI
}

