% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{class}
\name{IntersectBed_CLI-class}
\alias{IntersectBed_CLI-class}
\title{IntersectBed_CLI}
\description{
IntersectBed_CLI
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{inFilePath \code{"character"}}
   \item{\code{slot2}:}{inFileNames \code{"character"}}
   \item{\code{slot3}:}{cliParams \code{"character"}}
   \item{\code{slot4}:}{outFilePath \code{"character"}}
   \item{\code{slot5}:}{outputFlag \code{"character"}}
   \item{\code{slot6}:}{outFileName \code{"character"}}
 }
}

