% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{IntersectBed_CLI}
\alias{IntersectBed_CLI}
\title{Constructor method for IntersectBed_CLI}
\usage{
IntersectBed_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath,
  outFileName)
}
\arguments{
\item{inFilePath}{only one filename allowed!}
}
\description{
Constructor method for IntersectBed_CLI
}

