% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{class}
\name{MergeBedFile_CLI-class}
\alias{MergeBedFile_CLI-class}
\title{MergeBedFile_CLI}
\description{
MergeBedFile_CLI
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{inFilePath \code{"character"}}
   \item{\code{slot2}:}{inFileNames \code{"character"}}
   \item{\code{slot3}:}{cliParams \code{"character"}}
   \item{\code{slot4}:}{outFilePath \code{"character"}}
   \item{\code{slot5}:}{outputFlag \code{"character"}}
 }
}

