% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{class}
\name{MultiBamCov_CLI-class}
\alias{MultiBamCov_CLI-class}
\title{MultiBamCov_CLI}
\section{Slots}{
 
 \describe{
   \item{\code{slot1}:}{inFilePath \code{"character"}}
   \item{\code{slot2}:}{inFileNames \code{"character"}}
   \item{\code{slot3}:}{cliParams \code{"character"}}
   \item{\code{slot4}:}{outFilePath \code{"character"}}
   \item{\code{slot5}:}{outputFlag \code{"character"}}
   \item{\code{slot6}:}{annotationFileMB \code{"character"}}
   \item{\code{slot7}:}{annotationType \code{"character"}}
 }
}

