% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{MultiBamCov_CLI}
\alias{MultiBamCov_CLI}
\title{Constructor method for MultiBamCov_CLI}
\usage{
MultiBamCov_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath,
  annotationFileMB, annotationType)
}
\description{
Constructor method for MultiBamCov_CLI
}

