% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{class}
\name{MultiIntersectBed_CLI-class}
\alias{MultiIntersectBed_CLI-class}
\title{MultiIntersectBed_CLI}
\description{
MultiIntersectBed_CLI
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{inFilePath \code{"character"}}
   \item{\code{slot2}:}{inFileNames \code{"character"}}
   \item{\code{slot3}:}{cliParams \code{"character"}}
   \item{\code{slot4}:}{outFilePath \code{"character"}}
   \item{\code{slot5}:}{outputFlag \code{"character"}}
   \item{\code{slot6}:}{sortInFiles \code{"logical"}}
   \item{\code{slot7}:}{requireStrandness \code{"logical"}}
   \item{\code{slot8}:}{strandColumn \code{"numeric"}}
   \item{\code{slot9}:}{filterConditions \code{"character"}}
 }
}

