% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{class}
\name{MultiIntersectBed_perl_CLI-class}
\alias{MultiIntersectBed_perl_CLI-class}
\title{MultiIntersectBed_perl_CLI}
\description{
MultiIntersectBed_perl_CLI
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{inFilePath \code{"character"}}
   \item{\code{slot2}:}{inFileNames \code{"character"}}
   \item{\code{slot3}:}{cliParams \code{"character"}}
   \item{\code{slot4}:}{outFilePath \code{"character"}}
   \item{\code{slot5}:}{outputFlag \code{"character"}}
   \item{\code{slot6}:}{groupVect \code{"numeric"}}
   \item{\code{slot7}:}{outFileName \code{"character"}}
 }
}

