% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{MultiIntersectBed_perl_CLI}
\alias{MultiIntersectBed_perl_CLI}
\title{Constructor method for MultiIntersectBed_perl_CLI}
\usage{
MultiIntersectBed_perl_CLI(inFilePath, inFileNames, outputFlag, outFilePath,
  outFileName, ...)
}
\arguments{
\item{inFilePath}{(inFileNames may also be specified otherwise they will be fetched by list.files)}
}
\description{
Constructor method for MultiIntersectBed_perl_CLI
}

