% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{class}
\name{OutResultReference-class}
\alias{OutResultReference-class}
\title{OutResultReference Virtual Class}
\description{
OutResultReference Virtual Class
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{outResultName \code{"character"}}
 }
}

