% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{OutResultReference}
\alias{OutResultReference}
\title{Constructor (factory) for OutResultReference}
\usage{
OutResultReference(outResultName)
}
\value{
\code{OutResultReference} implementation
}

