% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{class}
\name{RNAStar_CLI-class}
\alias{RNAStar_CLI-class}
\title{RNAStar_CLI}
\section{Slots}{
 
 \describe{
   \item{\code{slot1}:}{inFilePath \code{"character"}}
   \item{\code{slot2}:}{inFileNames \code{"character"}}
   \item{\code{slot3}:}{cliParams \code{"character"}}
   \item{\code{slot4}:}{outFilePath \code{"character"}}
   \item{\code{slot5}:}{outputFlag \code{"character"}}
   \item{\code{slot6}:}{genomeIndexFilePath \code{"character"}}
   \item{\code{slot7}:}{filterSam \code{"logical"}}
   \item{\code{slot8}:}{outputFormat \code{"character"}}
 }
}

