% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{RNAStar_CLI}
\alias{RNAStar_CLI}
\title{Constructor method for RNAStar_CLI}
\usage{
RNAStar_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath,
  genomeIndexFilePath, filterSam, outputFormat)
}
\arguments{
\item{inFilePath}{(inFileNames may also be specified otherwise they will be fetched by list.files)}
}
\description{
Constructor method for RNAStar_CLI
}

