% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{class}
\name{Samtools_CLI-class}
\alias{Samtools_CLI-class}
\title{Samtools_CLI}
\description{
Samtools_CLI
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{inFilePath \code{"character"}}
   \item{\code{slot2}:}{inFileNames \code{"character"}}
   \item{\code{slot3}:}{cliParams \code{"character"}}
   \item{\code{slot4}:}{outFilePath \code{"character"}}
   \item{\code{slot5}:}{outputFlag \code{"character"}}
   \item{\code{slot6}:}{samtoolsApplication \code{"character"}}
   \item{\code{slot6}:}{outputFormat \code{"character"}}
 }
}

