% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{Samtools_CLI}
\alias{Samtools_CLI}
\title{Constructor method for Samtools_CLI}
\usage{
Samtools_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath,
  samtoolsApplication, outputFormat)
}
\arguments{
\item{inFilePath}{(inFileNames may also be specified otherwise they will be fetched by list.files)}
}
\description{
Constructor method for Samtools_CLI
}

