% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{class}
\name{Tophat2_CLI-class}
\alias{Tophat2_CLI-class}
\title{Tophat2_CLI}
\description{
Tophat2_CLI
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{inFilePath \code{"character"}}
   \item{\code{slot2}:}{inFileNames \code{"character"}}
   \item{\code{slot3}:}{cliParams \code{"character"}}
   \item{\code{slot4}:}{outFilePath \code{"character"}}
   \item{\code{slot5}:}{outputFlag \code{"character"}}
   \item{\code{slot6}:}{gtfFilePath \code{"character"}}
   \item{\code{slot7}:}{bowtieIndexFilePath \code{"character"}}
 }
}

