% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{Bowtie2_CLI}
\alias{Bowtie2_CLI}
\title{Constructor method for Bowtie2_CLI}
\usage{
Bowtie2_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath,
  bowtieIndexFilePath, matepairFileNames)
}
\arguments{
\item{inFilePath}{(inFileNames may also be specified otherwise they will be fetched by list.files)}
}
\description{
Constructor method for Bowtie2_CLI
}

