% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{class}
\name{CmdGenResult-class}
\alias{CmdGenResult-class}
\title{CmdGenResult}
\description{
CmdGenResult
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{CLIApplication \code{"CLIApplication"}}
   \item{\code{slot2}:}{OutResultReference \code{"OutResultReference"}}
   \item{\code{slot3}:}{commands \code{"character"}}
 }
}

