% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{CmdGenResult}
\alias{CmdGenResult}
\title{Constructor method for CmdGenResult}
\usage{
CmdGenResult(CLIApplication, OutResultReference, commands, executed)
}
\arguments{
\item{CLIApplication}{}

\item{OutResultReference}{}

\item{commands}{}

\item{executed}{}
}
\value{
\code{"CmdGenResult"}
}
\description{
Constructor method for CmdGenResult
}

