% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{CmdGenResult}
\alias{CmdGenResult}
\title{Constructor method for CmdGenResult}
\usage{
CmdGenResult(CLIApplication, OutResultReference, commands)
}
\arguments{
\item{CLIApplication}{}

\item{OutResultReference}{}

\item{commands}{}
}
\value{
\code{"CmdGenResult"}
}
\description{
Constructor method for CmdGenResult
}

