% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{class}
\name{CmdGenResultExec-class}
\alias{CmdGenResultExec-class}
\title{CmdGenResultExec}
\description{
CmdGenResultExec
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{cmdGenResult \code{"CmdGenResult"}}
   \item{\code{slot2}:}{execLog \code{"execLog"}}
 }
}

