% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{CmdGenResultExec}
\alias{CmdGenResultExec}
\title{Constructor method for CmdGenResultExec}
\usage{
CmdGenResultExec(cmdGenResult, execLog)
}
\arguments{
\item{execLog}{}

\item{CmdGenResult}{}
}
\value{
\code{"CmdGenResultExec"}
}
\description{
Constructor method for CmdGenResultExec
}

