% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{Cutadapt_CLI}
\alias{Cutadapt_CLI}
\title{Constructor method for Cutadapt_CLI}
\usage{
Cutadapt_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath)
}
\arguments{
\item{inFilePath}{(inFileNames may also be specified otherwise they will be fetched by list.files)}
}
\description{
Constructor method for Cutadapt_CLI
}

