% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{executeCommandResult}
\alias{executeCommandResult}
\title{Executes the commands of a executeCommandResult object}
\usage{
executeCommandResult(object)
}
\arguments{
\item{\code{"CmdGenResult"}}{}
}
\description{
Executes the commands of a executeCommandResult object
}

