% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{executeCommandResult}
\alias{executeCommandResult}
\title{Executes the commands of a executeCommandResult object}
\usage{
executeCommandResult(object, testing)
}
\arguments{
\item{\code{"CmdGenResult"}}{}
}
\description{
Executes the commands of a executeCommandResult object
}

