% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{FilesOutput}
\alias{FilesOutput}
\title{Constructor method for FilesOutput}
\usage{
FilesOutput(outResultName)
}
\arguments{
\item{outResultName}{}
}
\value{
\code{"FilesOutput"}
}
\description{
Constructor method for FilesOutput
}

