% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{FoldersOutput}
\alias{FoldersOutput}
\title{Constructor method for FoldersOutput}
\usage{
FoldersOutput(outResultName)
}
\arguments{
\item{outResultName}{}
}
\value{
\code{"FoldersOutput"}
}
\description{
Constructor method for FoldersOutput
}

