% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{generateCommandResult}
\alias{generateCommandResult}
\alias{generateCommandResult,Bowtie2TophatIon_CLI-method}
\alias{generateCommandResult,Bowtie2_CLI-method}
\alias{generateCommandResult,Cutadapt_CLI-method}
\alias{generateCommandResult,HTSeqCount_CLI-method}
\alias{generateCommandResult,MultiBamCov_CLI-method}
\alias{generateCommandResult,Samtools_CLI-method}
\alias{generateCommandResult,Tophat2_CLI-method}
\title{Generates the commands and possibly resulting files/folders}
\usage{
generateCommandResult(object)

\S4method{generateCommandResult}{Bowtie2_CLI}(object)

\S4method{generateCommandResult}{Bowtie2TophatIon_CLI}(object)

\S4method{generateCommandResult}{Cutadapt_CLI}(object)

\S4method{generateCommandResult}{HTSeqCount_CLI}(object)

\S4method{generateCommandResult}{MultiBamCov_CLI}(object)

\S4method{generateCommandResult}{Samtools_CLI}(object)

\S4method{generateCommandResult}{Tophat2_CLI}(object)
}
\arguments{
\item{CLIApplication}{}

\item{Bowtie2_CLI}{}

\item{Bowtie2TophatIon_CLI}{}

\item{Cutadapt_CLI}{}

\item{HTSeqCount_CLI}{}

\item{MultiBamCov_CLI}{}

\item{Samtools_CLI}{}

\item{Tophat2_CLI}{}
}
\value{
CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult

CmdGenResult
}
\description{
Generates the commands and possibly resulting files/folders

Generates the commands and possibly resulting files/folders of Bowtie2

Generates the commands and possibly resulting files/folders of Bowtie2 in combination with Tophat output for ion Torrent

Generates the commands and possibly resulting files/folders of Cutadapt

Generates the commands and possibly resulting files/folders of HTSeqCount_CLI

Generates the commands and possibly resulting files/folders of MultiBamCov_CLI

Generates the commands and possibly resulting files/folders of Samtools

Generates the commands and possibly resulting files/folders of Tophat2
}

