% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getBowtieIndexFilePath,Bowtie_CLI-method}
\alias{getBowtieIndexFilePath,Bowtie_CLI-method}
\title{Accessor getBowtieIndexFilePath}
\usage{
\S4method{getBowtieIndexFilePath}{Bowtie_CLI}(object)
}
\value{
bowtieIndexFilePath
}
\description{
Accessor getBowtieIndexFilePath
}

