% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getMatepairFileNames,Bowtie_CLI-method}
\alias{getMatepairFileNames,Bowtie_CLI-method}
\title{Accessor getMatepairFileNames}
\usage{
\S4method{getMatepairFileNames}{Bowtie_CLI}(object)
}
\value{
matepairFileNames
}

