% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getOutFileNames}
\alias{getOutFileNames}
\title{Accessor getOutFileNames}
\usage{
getOutFileNames(object)
}
\value{
outFileNames
}
\description{
Accessor getOutFileNames
}

