% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{getOutputFormat,BamToBed_CLI-method}
\alias{getOutputFormat,BamToBed_CLI-method}
\title{Accessor getOutputFormat}
\usage{
\S4method{getOutputFormat}{BamToBed_CLI}(object)
}
\value{
outputFormat
}
\description{
Accessor getOutputFormat
}

