% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{getOutputFormat,RNAStar_CLI-method}
\alias{getOutputFormat,RNAStar_CLI-method}
\title{Accessor getOutputFormat already defined in Samtools}
\usage{
\S4method{getOutputFormat}{RNAStar_CLI}(object)
}
\value{
outputFormat
}
\description{
Accessor getOutputFormat already defined in Samtools
}

