% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{HTSeqCount_CLI}
\alias{HTSeqCount_CLI}
\title{Constructor method for HTSeqCount_CLI}
\usage{
HTSeqCount_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath,
  gffFile)
}
\description{
Constructor method for HTSeqCount_CLI
}

