% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{IntersectBed_CLI}
\alias{IntersectBed_CLI}
\title{Constructor method for IntersectBed_CLI}
\usage{
IntersectBed_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath,
  outFileName)
}
\arguments{
\item{inFilePath}{only one filename allowed!}
}

