% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{MergeBedFile_CLI}
\alias{MergeBedFile_CLI}
\title{Constructor method for MergeBedFile_CLI}
\usage{
MergeBedFile_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath)
}
\arguments{
\item{inFilePath}{(inFileNames may also be specified otherwise they will be fetched by list.files)}
}
\description{
Constructor method for MergeBedFile_CLI
}

