% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{MultiBamCov_CLI}
\alias{MultiBamCov_CLI}
\title{Constructor method for MultiBamCov_CLI}
\usage{
MultiBamCov_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath,
  annotationFileMB, annotationType)
}

