% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{MultiIntersectBed_CLI}
\alias{MultiIntersectBed_CLI}
\title{Constructor method for MultiIntersectBed_CLI}
\usage{
MultiIntersectBed_CLI(inFilePath, inFileNames, cliParams, outputFlag,
  outFilePath, sortInFiles, requireStrandness, strandColumn, filterConditions)
}
\arguments{
\item{inFilePath}{(inFileNames may also be specified otherwise they will be fetched by list.files)}
}

