% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{class}
\name{OutResultReference-class}
\alias{OutResultReference-class}
\title{OutResultReference Virtual Class}
\description{
OutResultReference Virtual Class
}
\section{Slots}{

 \describe{
   \item{\code{slot1}:}{outResultName \code{"character"}}
 }
}

