% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{OutResultReference}
\alias{OutResultReference}
\title{Constructor (factory) for OutResultReference}
\usage{
OutResultReference(outResultName)
}
\arguments{
\item{outResultName}{}
}
\value{
\code{OutResultReference} implementation
}
\description{
Constructor (factory) for OutResultReference
}

