% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{OutResultReference}
\alias{OutResultReference}
\title{Constructor (factory) for OutResultReference}
\usage{
OutResultReference(outResultName)
}
\arguments{
\item{outResultName}{}
}
\value{
\code{OutResultReference} implementation
}
\description{
Constructor (factory) for OutResultReference
}

