% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singleRCodeFile.R
\docType{methods}
\name{SortFile_CLI}
\alias{SortFile_CLI}
\title{Constructor method for SortFile_CLI}
\usage{
SortFile_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath)
}
\arguments{
\item{inFilePath}{only one filename allowed!}
}
\description{
Constructor method for SortFile_CLI
}

