% Generated by roxygen2 (4.0.1.99): do not edit by hand
\docType{methods}
\name{Tophat2_CLI}
\alias{Tophat2_CLI}
\title{Constructor method for Tophat2_CLI}
\usage{
Tophat2_CLI(inFilePath, inFileNames, cliParams, outputFlag, outFilePath,
  bowtieIndexFilePath)
}
\arguments{
\item{inFilePath}{(inFileNames may also be specified otherwise they will be fetched by list.files)}
}
\description{
Constructor method for Tophat2_CLI
}

